Use byte instead of int where it makes sense too (MIDI bytes, bitmap, etc)

+++ src/us/rynet/akaifire/Screen.java
@@ -4,13 +4,13 @@ import us.rynet.akaifire.messages.ScreenMessage;
 
 public class Screen {
 
-  public int[][] BIT_MUTATE = { { 13, 19, 25, 31, 37, 43, 49 }, { 0, 20, 26, 32, 38, 44, 50 }, { 1, 7, 27, 33, 39, 45, 51 }, { 2, 8, 14, 34, 40, 46, 52 }, { 3, 9, 15, 21, 41, 47, 53 }, { 4, 10, 16, 22, 28, 48, 54 }, { 5, 11, 17, 23, 29, 35, 55 }, { 6, 12, 18, 24, 30, 36, 42 } };
+  public byte[][] BIT_MUTATE = { { 13, 19, 25, 31, 37, 43, 49 }, { 0, 20, 26, 32, 38, 44, 50 }, { 1, 7, 27, 33, 39, 45, 51 }, { 2, 8, 14, 34, 40, 46, 52 }, { 3, 9, 15, 21, 41, 47, 53 }, { 4, 10, 16, 22, 28, 48, 54 }, { 5, 11, 17, 23, 29, 35, 55 }, { 6, 12, 18, 24, 30, 36, 42 } };
 
-  protected int[] bitmap = new int[1171]; // The OLED has 1171 bytes for xfer: ceil(128*64/7)
+  protected byte[] bitmap = new byte[1171]; // The OLED has 1171 bytes for xfer: ceil(128*64/7)
 
-  public int[] getBitmap() { return bitmap; }
+  public byte[] getBitmap() { return bitmap; }
 
-  public void setBitmap(int[] bitmap) {
+  public void setBitmap(byte[] bitmap) {
     for (int x = 0; x < 128; x++) {
       for (int y = 0; y < 64; y++) {
         int index = x + (128 * y);
